accessible/tests/mochitest/treeupdate/test_canvas.html

<!DOCTYPE html>
<html>

<head>
  <title>Canvas subdom mutation</title>

  <link rel="stylesheet" type="text/css"
        href="chrome://mochikit/content/tests/SimpleTest/test.css" />

  <script src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>

  <script type="application/javascript"
          src="../common.js"></script>
  <script type="application/javascript"
          src="../role.js"></script>
  <script type="application/javascript"
          src="../events.js"></script>

  <script type="application/javascript">

    // //////////////////////////////////////////////////////////////////////////
    // Test

    async function doTest() {
      // make the subdom come alive!
      let shown = waitForEvent(EVENT_SHOW, getNode("dialog"));
      // ensure we start with no subtree
      testAccessibleTree("canvas", { CANVAS: [] });
      getNode("dialog").style.display = "block";
      await shown;
      testAccessibleTree("dialog", { DIALOG: [] });
      SimpleTest.finish();
    }

    SimpleTest.waitForExplicitFinish();
    addA11yLoadEvent(doTest);
  </script>
</head>
<body>

  <a target="_blank"
     title="Expose content in Canvas element"
     href="https://bugzilla.mozilla.org/show_bug.cgi?id=495912">
    Mozilla Bug 495912
  </a>

  <p id="display"></p>
  <div id="content" style="display: none"></div>
  <pre id="test">
  </pre>

  <canvas id="canvas">
    <div id="dialog" role="dialog" style="display: none;">
    </div>
  </canvas>

  <div id="eventdump"></div>
</body>
</html>
